\name{Convert Wind Velocity from km/h to m/s}
\alias{kmhtoms}
\title{Convert Wind Velocity from km/h to m/s
}
\description{
Converts Wind Velocity values in Km/h to m/s
}
\usage{
kmhtoms(wvkmh)
}
\arguments{
  \item{wvkmh}{Numerical Vector of Wind Velocity Values in Km/h
}
}
\value{Numerical vectors containing Wind velocity in m/s
}
\author{Francisco Jablinski Castelhano - Laboclima/Universidade Federal do Paraná
}
\examples{
##----Converting Wind from Km/h to m/s
kmhtoms(wvkmh=5)}
