\name{Convert Wind Velocity from knots to m/s}
\alias{kntoms}
\title{Convert Wind Velocity from knots to m/s
}
\description{
Converts Wind Velocity values in Knots to m/s
}
\usage{
kntoms(wvkn)
}
\arguments{
  \item{wvkn}{Numerical Vector of Wind Velocity Values in Knots
}
}
\value{Numerical vectors containing Wind velocity in m/s
}
\author{Francisco Jablinski Castelhano - Laboclima/Universidade Federal do Paraná
}
\examples{
##----Converting Wind from Knots to m/s
kntoms(wvkn=4)}
