\name{Convert Wind Velocity from mph to m/s}
\alias{mphtoms}
\title{Convert Wind Velocity from mph to m/s
}
\description{
Converts Wind Velocity values in mph to m/s
}
\usage{
mphtoms(wvmph)
}
\arguments{
  \item{wvmph}{Numerical Vector of Wind Velocity Values in mph
}
}
\value{Numerical vectors containing Wind velocity in m/s
}
\author{Francisco Jablinski Castelhano - Laboclima/Universidade Federal do Paraná}
\examples{
##----Converting Wind from mph to m/s
mphtoms(wvmph=9.4)}
