\name{Convert Temperature from Farenheit to Celsius}
\alias{tempftoc}
\title{Convert Temperature from Farenheit to Celsius
}
\description{Converts Temperature from Farenheit to Celsius}
\usage{
tempftoc(tf)
}
\arguments{
  \item{tf}{
Numerical Vector of Temperature Values in Farenheit
}
}

\value{Numerical vectors containing Temperature Values in Celsius
}
\author{Francisco Jablinski Castelhano - Laboclima/ Universidade Federal do Paraná
}
\examples{
##----Converting temperature from farenheit to celsius
tempftoc(tf= 92)}
